# Block3/api/objectsRequest/system.js
const prisma = require("../config/prisma");
const app = require("../config/app");

app.get("/api/system/all", async (req, res) => {
  // #swagger.tags = ['Systems']
  // #swagger.description = 'Получение всех систем'

  try {
    let systems = await prisma.system.findMany();
    res.json(systems);
  } catch (err) {
    console.error(err);
    res.status(500).send(err.message);
  }
});

app.post("/api/system/create", async (req, res) => {
  // #swagger.tags = ['Systems']
  // #swagger.description = 'Получение всех систем'

  try {
    let check = await prisma.system.findUnique({
      where: { Key: String(req.body["Key"]) },
    });
    if (check) {
      res.status(400).send("Key is not unique");
      return;
    }
    let systems = await prisma.system.create({
      data: {
        Name: req.body["Name"],
        Key: String(req.body["Key"]),
      },
    });
    res.json(systems);
  } catch (err) {
    console.error(err);
    res.status(500).send(err.message);
  }
});