# Block3/api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.2",
    "title": "Notify",
    "description": "API for notify service"
  },
  "host": "localhost:3080",
  "basePath": "/",
...
  "schemes": [
    "https"
  ],
  "securityDefinitions": {
    "AccessToken": {
      "type": "apiKey",
      "name": "x-access-token",
      "in": "header"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/system/create": {
      "post": {
        "tags": [
          "Systems"
        ],
        "description": "Получение всех систем",
        "parameters": [],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
...
  "security": [
    {
      "AccessToken": []
    }
  ]
}